Modifying run routine.

+++ docker/bin/init.sh
@@ -10,7 +10,7 @@ quick_iam_test() {
     printf "\033[93m>\033[0m Testing IAM with '%s' ...\n" "user"
     printf "\033[96;1m%s\033[0m\n" "aws-vault exec user -- aws sts get-caller-identity"
     aws-vault exec user -- aws sts get-caller-identity
-    
+
     sleep 1
     printf "\033[93m>\033[0m Testing IAM with '%s' ...\n" "${team}-${cluster}"
     printf "\033[96;1m%s\033[0m\n" "aws-vault exec ${team}-${cluster} -- aws sts get-caller-identity"
@@ -63,35 +63,39 @@ init_aws() {
     [ $last_err -eq 0 -a -f /.initialized ] && printf "\033[92;1m<<< Successfully Initialized %s <<<\033[0m\n" "AWS (and dpctl)"
 }
 
-tarballs="$(find /tmp/addons -mindepth 1 -type f -name '*.tgz' | grep -v /archive)"
-if [ $(count ${tarballs[@]}) -gt 0 ]; then
-    [ -d /tmp/addons/archive ] || mkdir /tmp/addons/archive
-    pushd /tmp/addons
-    for f in ${tarballs[@]}; do tar -xzvf "${f}"; done
-    popd
-    mv ${tarballs[@]} /tmp/addons/archive/ 2>/dev/null || true
-fi
+if [ -d /tmp/addons ]; then
+    tarballs="$(find /tmp/addons -mindepth 1 -type f -name '*.tgz' | grep -v /archive)"
+    if [ $(count ${tarballs[@]}) -gt 0 ]; then
+        [ -d /tmp/addons/archive ] || mkdir /tmp/addons/archive
+        pushd /tmp/addons
+        for f in ${tarballs[@]}; do tar -xzvf "${f}"; done
+        popd
+        mv ${tarballs[@]} /tmp/addons/archive/ 2>/dev/null || true
+    fi
 
-zips="$(find /tmp/addons -mindepth 1 -type f -name '*.zip' | grep -v /archive)"
-if [ $(count ${zips[@]}) -gt 0 ]; then
-    [ -d /tmp/addons/archive ] || mkdir /tmp/addons/archive
-    pushd /tmp/addons
-    for f in ${zips[@]}; do unzip -o "${f}"; done
-    popd
-    mv ${zips[@]} /tmp/addons/archive/ 2>/dev/null || true
-fi
+    zips="$(find /tmp/addons -mindepth 1 -type f -name '*.zip' | grep -v /archive)"
+    if [ $(count ${zips[@]}) -gt 0 ]; then
+        [ -d /tmp/addons/archive ] || mkdir /tmp/addons/archive
+        pushd /tmp/addons
+        for f in ${zips[@]}; do unzip -o "${f}"; done
+        popd
+        mv ${zips[@]} /tmp/addons/archive/ 2>/dev/null || true
+    fi
 
-files="$(find /tmp/addons -mindepth 1 -type f | grep -v -P '\.tgz|\.zip|/archive')"
-if [ $(count ${files[@]}) -gt 0 ]; then
-    for f in ${files[@]}; do
-        fname="$(basename ${f})"
-        target="/usr/local/bin/${fname}"
-        if [ ! -f "${target}" ]; then
-            printf "\033[93m>\033[0m Installing '%s' to '%s'\n" "${f}" "/usr/local/bin/${fname}"
-            install "${f}" "${target}"
-        fi
-    done
-    rm -f ${files[@]}
+    files="$(find /tmp/addons -mindepth 1 -type f | grep -v -P '\.tgz|\.zip|/archive')"
+    if [ $(count ${files[@]}) -gt 0 ]; then
+        for f in ${files[@]}; do
+            fname="$(basename ${f})"
+            target="/usr/local/bin/${fname}"
+            if [ ! -f "${target}" ]; then
+                printf "\033[93m>\033[0m Installing '%s' to '%s'\n" "${f}" "/usr/local/bin/${fname}"
+                install "${f}" "${target}"
+            fi
+        done
+        rm -f ${files[@]}
+    fi
 fi
 
 [ -z "$(which dpctl)" ] || init_aws
+
+export -f count
